+++ bot/src/main/java/by/ourgame/bot/repository/GameRepository.java
@@ -8,9 +8,9 @@ import reactor.core.publisher.Mono;
 @Repository
 public interface GameRepository extends ReactiveMongoRepository<Game, String> {
 
-    Mono<Game> findByChat_Id(Integer id);
+    Mono<Game> findByChat_Id(Long id);
 
-    Mono<Game> findByCreator_Id(Integer id);
+    Mono<Game> findByCreator_Id(Long id);
 
-    Mono<Game> findByChat_IdAndCanAnswer(Integer id, boolean canAnswer);
+    Mono<Game> findByChat_IdAndCanAnswer(Long id, boolean canAnswer);
 }
